Use httpx mounts instead of the deprecated proxies
As of httpx v 0.28 the deprecated proxies option has been removed. Use the
mounts option instead.  This fix is minimal in the sense that it just
before the call to httpx switches out the proxies argument for a mounts
argument.  This could probably be more nicely solved at lower level, but
this seems to fix the problem with minimal changes.